Basic animations are in 1- Cards move from the draw deck to the board and flip over 2- When new cards are drawn, the board fades to show resized card, then 3 new cards are moved onto the board and flip 3- When cards are selected or unselected, the border fades in and out 4- When cards are matched, they move to the discard pile and new drawn cards move to their old spaces and flip
Modified the game's card replacement algorithm to support talking to the view controller, removed all private modifiers.  Will reassess private member elements after we merge back to the trunk

these animations need to be debugged and we need to handle the end of the game cases when the number of cards on the board decreases.